Refactor cognitive_folio settings and services
- Removed the 'tool_calls_enabled' setting from cf_settings.js and cf_settings.json.
- Introduced new service files: PromptProcessor, SettingsManager, StreamingHandler, TokenManager, ToolOrchestrator, and WebSearchService.
- Implemented methods for prompt processing, settings management, streaming response handling, token management, tool orchestration, and web searching.
- Enhanced error handling and logging for web search and tool execution processes.
- Improved validation for chat settings in SettingsManager.
