+++ cognitive_folio/cognitive_folio/doctype/cf_settings/cf_settings.js
@@ -31,26 +31,14 @@ function ensure_runtime_defaults(frm) {
         retry_backoff_base_seconds: 1.5,
         stream_flush_interval_seconds: 1.0,
         stream_flush_min_char_delta: 120,
-        tool_calls_enabled: 1,
         max_tool_rounds: 8,
         max_tool_calls_per_round: 8,
         tool_result_max_chars: 8000,
     };
 
-    const checkboxDefaults = ['tool_calls_enabled'];
-
     let changed = false;
 
     Object.keys(defaults).forEach((fieldname) => {
-        if (checkboxDefaults.includes(fieldname)) {
-            const currentValue = frm.doc[fieldname];
-            if (currentValue === undefined || currentValue === null || currentValue === '') {
-                frm.set_value(fieldname, defaults[fieldname]);
-                changed = true;
-            }
-            return;
-        }
-
         const currentValue = Number(frm.doc[fieldname] || 0);
         if (!(currentValue > 0)) {
             frm.set_value(fieldname, defaults[fieldname]);
